Update patch set 1
Patch Set 1:

yeah, that's something we were pretty confused on. The build would fail until we moved the ion.h into /inc. The "" vs <> was also a confusion. I can switch that back when I get home. the hal_public.h also caused a make error since GRALLOC_BUFFER was defined in Decoder.c Once we moved that to /inc, the build completed.
I did notice it's the same header as what's in the /hwc source. Is this even needed in the decoder source?

Patch-set: 1