Update patch set 2
Patch Set 2: Code-Review-1

I'd be inclined to say it's a mistake by LG.

This code is included in LG's v30c source, but not the v30d release.  There is indeed a NOTICE file containing a copy of the LGPL in hardware/ti/omap3/omx/video/src/openmax_il -- but there are also separate NOTICE files under each of the components there that we expect to be open-source, while the Ittiam components have no NOTICE.

Meanwhile, not only are the license headers for the Ittiam components all wrong, the "source" includes binary blobs which are linked into the final product.  It's unlikely that anyone would release binary blobs under LGPL (there's a school of thought that says you can't redistribute binary blobs released by the copyright holder under GPL/LGPL: the license only grants redistribution rights for "source code", defined as "the preferred form of the work for making modifications to it", which the blobs clearly aren't), and I don't see any separate license terms for the blobs.

If this is indeed a mistake by LG, then the fact that it's out there doesn't mean that we can distribute it -- only the copyright holder (Ittiam) can grant that permission.

Patch-set: 2
Label: Code-Review=-1